+++ apps/bridge/components/layerZero/config.ts
@@ -6,3 +6,7 @@ export const LAYER_ZERO_JS = {
   css: `https://unpkg.com/@layerzerolabs/x-pancakeswap-widget@${VERSION}/element.css`,
   integrity: `sha384-${SHA384}`,
 }
+
+export const PARTNER_ID = 0x0002
+export const FEE_COLLECTOR = '0x68C7ABB8b1c3D1cE467E28265770F3a7ECF32654'
+export const FEE_TENTH_BPS = '0'
